Moved DataValue to data.value
