+++ src/main/java/model/data/DataModelBuilder.java
@@ -1,6 +1,10 @@
 package model.data;
 
 
+import exceptions.ColumnValueMismatchException;
+import exceptions.ColumnValueTypeMismatchException;
+import model.data.value.DataValue;
+
 import java.util.ArrayList;
 import java.util.List;
