+++ src/main/java/model/data/DataRow.java
@@ -2,6 +2,7 @@ package model.data;
 
 import exceptions.ColumnValueMismatchException;
 import exceptions.ColumnValueTypeMismatchException;
+import model.data.value.DataValue;
 
 import java.util.HashMap;
 import java.util.Map;
