+++ src/main/java/model/data/describer/ConstantDescriber.java
@@ -1,7 +1,7 @@
 package model.data.describer;
 
 import model.data.DataRow;
-import model.data.DataValue;
+import model.data.value.DataValue;
 
 /**
  * The describer for values that don't change.
