+++ src/main/java/model/data/describer/RowValueDescriber.java
@@ -2,7 +2,7 @@ package model.data.describer;
 
 import model.data.DataColumn;
 import model.data.DataRow;
-import model.data.DataValue;
+import model.data.value.DataValue;
 
 /**
  * A describer for values contained within a table.
